requirements.txt
django == 5.1.3
django-rest-framework
django-crispy-forms
crispy-bootstrap5
reportlab